+++ server/package.json
@@ -37,7 +37,7 @@
   },
   "devDependencies": {
     "@types/fs-extra": "^8.1.0",
-    "npm-groovy-lint": "^3.2.2",
+    "npm-groovy-lint": "^3.2.3",
     "vscode-languageserver-textdocument": "^1.0.1"
   }
 }
